change datatypes on db to be bigints

+++ twitoff/models.py
@@ -6,14 +6,14 @@ db = SQLAlchemy(session_options={"autoflush": False})
 migrate = Migrate()
 
 class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     screen_name = db.Column(db.String(128))
     followers_count = db.Column(db.Integer)
     latest_tweet_id = db.Column(db.BigInteger)
 
 class Tweet(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"))
+    id = db.Column(db.BigInteger, primary_key=True)
+    user_id = db.Column(db.BigInteger, db.ForeignKey("user.id"))
     full_text = db.Column(db.String(500))
     embedding = db.Column(db.PickleType)
